move UsersList file to diff folder
